Change build system to support multiple Linux images
Adds a new Kconfig variable under seL4 Applications ---> Linux
that gives a prefix for the image files in apps/linux/(platform). This
prefix will be replaced by linux- when the images are copied.
Defaults to Image for zynqmp and fsl-image-vm for imx8.

Also changes the configuration for zynqmp and imx8 to look for
an image named linux-1

Change-Id: I1f9fc38e193c77cf48cd7f7a103698ee5fed49c4